v0.8 Remove empty fields

+++ backend/call_cli.py
@@ -21,17 +21,25 @@ def main():
             if params == job_address:
                 file_path = Path().joinpath("tmp").joinpath(job_address + ".txt")
                 with open(file_path) as f:
-                    params = json.dumps(json.load(f))
+                    params = json.load(f)
+            else:
+                params = json.loads(params)
+
+            tmp_params = {}
+            for k,v in params.items():
+                if v != "":
+                    tmp_params[k] = v
+            params = json.dumps(tmp_params)
 
             print(call(job_address, job_signature, endpoint, spec_hash, method, params))
 
-            if os.path.exists(file_path):
+            if file_path and os.path.exists(file_path):
                 os.remove(str(file_path))
         else:
             print("Invalid number of Params!")
     except Exception as e:
         print(e)
-        if os.path.exists(file_path):
+        if file_path and os.path.exists(file_path):
             os.remove(str(file_path))
         return
